Refactor operation code
