+++ projects/optic/src/commands/capture/patches/patchers/closeness/schema-inventory.ts
@@ -1,10 +1,10 @@
 import { FlatOpenAPIV3, OpenAPIV3 } from '@useoptic/openapi-utilities';
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
 import { computeClosenessFromKeyValueTuples, walkSchema } from './closeness';
-import { SpecPatch, SpecPatches } from '../../../../oas/specs';
 import { PatchImpact } from '../../patch-operations';
 import { PathComponents } from '../../../../oas/operations';
 import { Operation } from 'fast-json-patch';
+import { SpecPatch, SpecPatches } from '../spec/patches';
 
 type ClosestMatch = { ref: string; percent: number } | undefined;
 export class SchemaInventory {
