+++ projects/optic/src/commands/capture/patches/patchers/shapes/handlers/oneOf.ts
@@ -8,7 +8,8 @@ import {
   ShapeDiffResultKind,
 } from '../diff';
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
-import { OperationGroup, PatchImpact } from '../../../../../oas/specs/patches';
+import { PatchImpact } from '../../../patch-operations';
+import { OperationGroup } from '../../spec/patches';
 import { ShapePatch } from '../patches';
 import { CapturedInteraction } from '../../../../sources/captured-interactions';
