+++ projects/optic/src/commands/capture/patches/patchers/shapes/handlers/required.ts
@@ -6,7 +6,8 @@ import {
   ShapeDiffResultKind,
 } from '../diff';
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
-import { OperationGroup, PatchImpact } from '../../../../../oas/specs/patches';
+import { PatchImpact } from '../../../patch-operations';
+import { OperationGroup } from '../../spec/patches';
 import { SchemaObject } from 'ajv';
 import { ShapeLocation } from '../documented-bodies';
 import { ShapePatch } from '../patches';
