+++ projects/optic/src/commands/oas/operations/diffs/index.ts
@@ -1,23 +1,9 @@
-import { Operation } from '..';
-import { CapturedInteraction } from '../../../capture/sources/captured-interactions';
 import { OpenAPIV3 } from '../../specs';
 import { OperationDiffResult, OperationDiffResultKind } from './result';
-import {
-  OperationInteractionDiffTraverser,
-  SpecOperationDiffTraverser,
-} from './traversers';
+import { SpecOperationDiffTraverser } from './traversers';
 
 export type { OperationDiffResult };
 export { OperationDiffResultKind };
-
-export function* diffInteractionByOperation(
-  interaction: CapturedInteraction,
-  operation: Operation
-): IterableIterator<OperationDiffResult> {
-  const traverser = new OperationInteractionDiffTraverser();
-  traverser.traverse(interaction, operation);
-  yield* traverser.results();
-}
 
 export function* diffOperationWithSpec(
   operation: {
