+++ projects/optic/src/commands/oas/operations/diffs/traversers.ts
@@ -1,34 +1,11 @@
-import { Operation } from '..';
 import { OperationDiffResult } from './result';
 import {
   OperationQueries,
   specToOperations,
 } from '../../../capture/operations/queries';
 import { OpenAPIV3 } from '../../specs';
-import { visitRequestBody, visitResponses } from './visitors/index';
 import { visitPath, visitMethod } from './visitors/index';
 import { Some, None } from 'ts-results';
-import { CapturedInteraction } from '../../../capture/sources/captured-interactions';
-
-export class OperationInteractionDiffTraverser {
-  private interaction?: CapturedInteraction;
-  private operation?: Operation;
-
-  traverse(interaction, operation) {
-    this.interaction = interaction;
-    this.operation = operation;
-  }
-
-  *results(): IterableIterator<OperationDiffResult> {
-    if (!this.interaction || !this.operation) return;
-
-    const { operation, interaction } = this;
-
-    yield* visitRequestBody(interaction.request, operation.requestBody);
-
-    yield* visitResponses(interaction.response, operation.responses);
-  }
-}
 
 export class SpecOperationDiffTraverser {
   private operation?: {
