+++ projects/optic/src/commands/oas/specs/index.ts
@@ -11,8 +11,6 @@ export { SpecFile, SpecFilesSourcemap } from './files';
 export type { SpecFileOperation } from './files';
 
 // patches and operations
-export { SpecPatch } from './patches';
-export { SpecPatches } from './streams/patches';
 export { SpecFileOperations, SpecFiles, SpecFilesAsync } from './streams/files';
 export {
   SpecFacts,
